+++ src/impls/core/cmp.rs
@@ -1,9 +1,10 @@
 use crate::{
-    typeset::{tset, ComputeTypeSet, ExpandTypeHList},
-    Field, MaybeUninhabited, TypeLayout, TypeLayoutInfo, TypeStructure, Variant,
+    graph::hlist, Field, MaybeUninhabited, TypeLayout, TypeLayoutInfo, TypeStructure, Variant,
 };
 
 unsafe impl<T: TypeLayout> TypeLayout for core::cmp::Reverse<T> {
+    type TypeGraphEdges = hlist![T];
+
     const INHABITED: crate::MaybeUninhabited = T::INHABITED;
     const TYPE_LAYOUT: TypeLayoutInfo<'static> = TypeLayoutInfo {
         name: ::core::any::type_name::<Self>(),
@@ -20,11 +21,9 @@ unsafe impl<T: TypeLayout> TypeLayout for core::cmp::Reverse<T> {
     };
 }
 
-unsafe impl<T: ComputeTypeSet> ComputeTypeSet for core::cmp::Reverse<T> {
-    type Output<R: ExpandTypeHList> = tset![T, .. @ R];
-}
-
 unsafe impl TypeLayout for core::cmp::Ordering {
+    type TypeGraphEdges = hlist![::core::mem::Discriminant<Self>];
+
     const INHABITED: crate::MaybeUninhabited = crate::inhabited::all![];
     const TYPE_LAYOUT: TypeLayoutInfo<'static> = TypeLayoutInfo {
         name: ::core::any::type_name::<Self>(),
@@ -51,10 +50,4 @@ unsafe impl TypeLayout for core::cmp::Ordering {
             ],
         },
     };
-}
-
-unsafe impl ComputeTypeSet for core::cmp::Ordering {
-    type Output<R: ExpandTypeHList> = tset![
-        ::core::mem::Discriminant<Self>, .. @ R
-    ];
 }
