+++ src/impls/core/ref.rs
@@ -1,9 +1,8 @@
-use crate::{
-    typeset::{tset, ComputeTypeSet, ExpandTypeHList},
-    TypeLayout, TypeLayoutInfo, TypeStructure,
-};
+use crate::{graph::hlist, TypeLayout, TypeLayoutInfo, TypeStructure};
 
 unsafe impl<'a, T: TypeLayout + 'a> TypeLayout for &'a T {
+    type TypeGraphEdges = hlist![T];
+
     const INHABITED: crate::MaybeUninhabited = crate::inhabited::all![];
     const TYPE_LAYOUT: TypeLayoutInfo<'static> = TypeLayoutInfo {
         name: ::core::any::type_name::<Self>(),
@@ -11,13 +10,11 @@ unsafe impl<'a, T: TypeLayout + 'a> TypeLayout for &'a T {
         alignment: ::core::mem::align_of::<Self>(),
         structure: TypeStructure::Primitive,
     };
-}
-
-unsafe impl<'a, T: ComputeTypeSet + 'a> ComputeTypeSet for &'a T {
-    type Output<R: ExpandTypeHList> = tset![T, .. @ R];
 }
 
 unsafe impl<'a, T: TypeLayout + 'a> TypeLayout for &'a mut T {
+    type TypeGraphEdges = hlist![T];
+
     const INHABITED: crate::MaybeUninhabited = crate::inhabited::all![];
     const TYPE_LAYOUT: TypeLayoutInfo<'static> = TypeLayoutInfo {
         name: ::core::any::type_name::<Self>(),
@@ -25,8 +22,4 @@ unsafe impl<'a, T: TypeLayout + 'a> TypeLayout for &'a mut T {
         alignment: ::core::mem::align_of::<Self>(),
         structure: TypeStructure::Primitive,
     };
-}
-
-unsafe impl<'a, T: ComputeTypeSet + 'a> ComputeTypeSet for &'a mut T {
-    type Output<R: ExpandTypeHList> = tset![T, .. @ R];
 }
