+++ src/impls/core/sync/atomic.rs
@@ -1,7 +1,4 @@
-use crate::{
-    typeset::{tset, ComputeTypeSet, ExpandTypeHList},
-    Field, MaybeUninhabited, TypeLayout, TypeLayoutInfo, TypeStructure,
-};
+use crate::{graph::hlist, Field, MaybeUninhabited, TypeLayout, TypeLayoutInfo, TypeStructure};
 
 macro_rules! impl_atomic_int_layout {
     (impl $at:ident ( $align:literal : $cfg:literal ) => $ty:ty => $val:literal) => {
@@ -24,11 +21,8 @@ macro_rules! impl_atomic_int_layout {
                     ],
                 },
             };
-        }
 
-        #[cfg(target_has_atomic_load_store = $cfg)]
-        unsafe impl ComputeTypeSet for core::sync::atomic::$at {
-            type Output<T: ExpandTypeHList> = tset![core::cell::UnsafeCell<$ty>, .. @ T];
+            type TypeGraphEdges = hlist![core::cell::UnsafeCell<$ty>];
         }
     };
     ($($at:ident ( $align:literal : $cfg:literal ) => $ty:ty => $val:literal),*) => {
@@ -71,12 +65,8 @@ macro_rules! impl_atomic_int_ptr_sized_layout {
                     ],
                 },
             };
-        }
 
-        #[cfg(target_has_atomic_load_store = "ptr")]
-        #[cfg(target_pointer_width = $cfg)]
-        unsafe impl ComputeTypeSet for core::sync::atomic::$at {
-            type Output<T: ExpandTypeHList> = tset![core::cell::UnsafeCell<$ty>, .. @ T];
+            type TypeGraphEdges = hlist![core::cell::UnsafeCell<$ty>];
         }
     };
     ($($at:ident ( $align:literal : $cfg:literal ) => $ty:ty => $val:literal),*) => {
@@ -113,12 +103,8 @@ macro_rules! impl_atomic_ptr_layout {
                     ],
                 },
             };
-        }
 
-        #[cfg(target_has_atomic_load_store = "ptr")]
-        #[cfg(target_pointer_width = $cfg)]
-        unsafe impl<T: ComputeTypeSet> ComputeTypeSet for core::sync::atomic::AtomicPtr<T> {
-            type Output<R: ExpandTypeHList> = tset![core::cell::UnsafeCell<T>, .. @ R];
+            type TypeGraphEdges = hlist![core::cell::UnsafeCell<T>];
         }
     };
     ($(( $align:literal : $cfg:literal )),*) => {
